requirements.txt
fastapi
granian[rloop,reload,dotenv]
python-dotenv
rloop
colorama
scalar-fastapi
renoir
loguru
orjson
fastapi-limiter
redis[hiredis]
pydantic
pydantic-settings